Big commit with all modifications over the 2 last month
