+++ silx/opencl/sift/match.py
@@ -38,8 +38,11 @@ __authors__ = ["Jérôme Kieffer", "Pierre Paleo"]
 __contact__ = "jerome.kieffer@esrf.eu"
 __license__ = "MIT"
 __copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "12/01/2017"
+__date__ = "15/03/2017"
 __status__ = "production"
+
+
+import os
 import logging
 import threading
 import gc
@@ -174,7 +177,7 @@ class MatchPlan(object):
         for kernel in list(self.kernels.keys()):
             if "." in kernel: 
                 continue
-            kernel_src = get_opencl_code(kernel)
+            kernel_src = get_opencl_code(os.path.join("sift", kernel))
 
             wg_size = self.kernels[kernel]
             try:
