+++ silx/opencl/sift/test/test_convol.py
@@ -37,8 +37,9 @@ __authors__ = ["Jérôme Kieffer", "Pierre Paleo"]
 __contact__ = "jerome.kieffer@esrf.eu"
 __license__ = "MIT"
 __copyright__ = "2013 European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "12/01/2017"
+__date__ = "15/03/2017"
 
+import os
 import time
 import logging
 import numpy
@@ -106,7 +107,7 @@ class TestConvol(unittest.TestCase):
         self.gpu_in = pyopencl.array.to_device(self.queue, self.input)
         self.gpu_tmp = pyopencl.array.empty(self.queue, self.input.shape, dtype=numpy.float32, order="C")
         self.gpu_out = pyopencl.array.empty(self.queue, self.input.shape, dtype=numpy.float32, order="C")
-        kernel_src = get_opencl_code("convolution.cl")
+        kernel_src = get_opencl_code(os.path.join("sift", "convolution.cl"))
         self.program = pyopencl.Program(self.ctx, kernel_src).build()
         self.IMAGE_W = numpy.int32(self.input.shape[-1])
         self.IMAGE_H = numpy.int32(self.input.shape[0])
