+++ silx/opencl/sift/test/test_reductions.py
@@ -37,7 +37,7 @@ __authors__ = ["Jérôme Kieffer", "Pierre Paleo"]
 __contact__ = "jerome.kieffer@esrf.eu"
 __license__ = "MIT"
 __copyright__ = "2013 European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "12/01/2017"
+__date__ = "15/03/2017"
 
 import os
 import unittest
@@ -87,7 +87,7 @@ class TestReduction(unittest.TestCase):
         cls.queue = None
 
     def setUp(self):
-        kernel_src = get_opencl_code("reductions")
+        kernel_src = get_opencl_code(os.path.join("sift", "reductions"))
         self.program = pyopencl.Program(self.ctx, kernel_src).build(options="")
 
     def tearDown(self):
